Male Sankey with people nodes
Added people nodes to the sankey diagram (male-based lineage).  It
now shows people branching (hyperedges) from a birth-marriage through
the person node and out to all marriages in which that person was
married as an adult.  Highlights work over the person node.  All text
has been moved into the hovers for a cleaner diagram.