Better search functionality
Added a search functionality that allows for searching from the index
page.  Also created a way to click people in the search list to get
their children.  In the normal people view and marriage view, there is a
link to switch back and forth to see parents and children of a marriage.
Queries in the people and marriage tables now only load those that match
the search so that page load times are faster.