Brown Data in Data Entry Page
Added APIs to access the brown data, both in its entirety and per
individual.  This allows us to get the list of people in Brown and the
Brown information for a particular individual.

Added the list of people to the data_entry view with an Edit button to
take the user to that person's edit page.

Added Brown information to the side of the edit page so that the user
can reference it while editing an entry.