Attached Temple Rites
Added correct SQL to grab temple rite information to the edit_person
api, and put the correct hooks in the individual edit page.  Also
attached the person search to the person selects in the individual edit
page by having the scripts js file use the get_person api to search for
people.