Working database save
The save action now updates the nauvoo_data_test database correctly.
Updates happen if it is possible to do an update, else it does an
insert.   All changes are logged to "submissions.txt" as a legal SQL
dump file with comments indicating the timestamp.  Any errors are listed
in comments directly below their SQL line.