Fixed Multiple Save Creation Issue
After adding a new entry to any part, clicking save multiple times would
result in adding a new entry to the database for every save.  Now, the
save script returns the new IDs in a JSON object with mappings to the
save page's IDs.  On a success, the save javascript now takes those
returned ids and replaces the NEW values in the save page with the
correct IDs.