Marriage Bug Fix
There was an assumption that marriages always had a husband and a wife.
Thus, they were right joins of PersonMarriage to Marriage when grabbing
all the information about a person.  This might not be the case, if
currently editing data.  So, they have been changed to Left Joins,
allowing for one or both participants to be null and allowing a person
to see and edit that marriage.