Fixed Editing and Inserting Marriages
Before, I had used "update" and if not "insert".  However, the update
returns that it truly did something, even if it doesn't do anything.
This meant that if a person was not linked to a marraige (either a new
marriage or adding a person to an existing one), then the person would
never get inserted or linked.  Now, the people are checked a little more
thoroughly.  A select is done.

If the role and marriage exist, then: if the personid exists, do an
update, else do a delete for that marriage/role pair in PersonMarriage.

If the role and marriage don't exist and the personid does, then insert
this new person/role for this marriageid in the PersonMarriage table.