Ability to create individuals
Added the functionality in the edit page and supporting scripts to allow
the creation of a new individual.  If the edit page gets ?id=NEW, then
it will query the api with NEW which will return an empty person
structure to fill in.  When the first save happens on the edit page, it
will then insert a new Person entry in the database and update the IDs
on the edit page with the correct newly created id.