Separated Offices
Offices are now their own table.  Now, the database contains an Office
table for the offices and a PersonOffice join table that connects a
person to their offices.  The data_entry page now has a new tab for
offices, and the offices may be added and removed in that manner.